# Only need to throw once in cast expr