Summary:
This is a performance optimization

before:
```
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast_int##try_invalid                                        4.88s   204.83m
cast_int##try_valid                                         2.15ms    465.22
cast_int##valid                                             2.20ms    453.96
----------------------------------------------------------------------------
```

after:
```

============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast_int##try_invalid                                        3.05s   328.35m
cast_int##try_valid                                         2.24ms    445.58
cast_int##valid                                             2.30ms    433.92
----------------------------------------------------------------------------
```

Differential Revision: D47338098